Player: Added an 'addItem' function which adds the passed item to the player's inventory.
MainListener: Added a 'pickup' function which will pick up a valid Item from a Room. It also removes said item from the Room so that nobody else can pick it up.

Initializer: Added a Lamp to the 'aHouse' Room.

Room: Added a 'searchItem' function which takes in a String and compares it to the names of all items in the current Room.